# model/comment_model.go
package model

import "time"

type MComment struct {
	ID        int64     `json:"id"`
	PostID    int64     `json:"post_id"`
	UserID    int64     `json:"user_id"`
	Content   string    `json:"content"`
	CreatedAt time.Time `json:"created_at"`
}

type CreateCommentReq struct {
	Content string `json:"content" binding:"required"`
}